requirements.txt
RPi.GPIO
pigpio
getch

